Move announcement processing code into a separate thread refactored into a separate function
